♻️ Refacto footer
