--- components/layout/footer/Address.vue
@@ -1,11 +1,0 @@
-<template>
-  <a
-    href="https://goo.gl/maps/aScUymM4T4DqeBvz8"
-    target="_blank"
-    aria-label="lien vers le palais des congrès sur google maps - Nouvelle fenêtre"
-    class="block"
-  >
-    <span class="font-bold text-2xl">Palais des Congrès</span><br>
-    <span>Avenue Jean Gabriel Domergue<br>33300 BORDEAUX</span>
-  </a>
-</template>
